% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulation.R
\name{GetPopulation}
\alias{GetPopulation}
\title{GetPopulation function
This function reads the population data from the EURO dashboard}
\usage{
GetPopulation()
}
\description{
GetPopulation function
This function reads the population data from the EURO dashboard
}
